+++ e2e/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @reaatech/mcp-contract-e2e
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @reaatech/mcp-contract-cli@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes